% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_features.R
\name{bc_features}
\alias{bc_features}
\title{Compute features for the TSP instance using Concorde algorithm.}
\usage{
bc_features(path_concorde, instance_path)
}
\arguments{
\item{path_concorde}{Path to the Concorde algorithm executable.}

\item{instance_path}{Path to the TSP instance file.}
}
\value{
A vector containing computed features.
}
\description{
This function combines multiple features computed using Concorde algorithm for a TSP instance.
}
\examples{
\dontrun{
  bc_features("/path/to/concorde", "/path/to/tsp_instance")
}

}
