% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{callAustralianSolverInterface}
\alias{callAustralianSolverInterface}
\title{Adapter/interface to the TSP algorithms implemented by the group of Markus Wagner.}
\usage{
callAustralianSolverInterface(instance, control, solver, bin)
}
\arguments{
\item{instance}{[\code{character(1)}]\cr
Path to instance in TSPlib format.}

\item{control}{[\code{list}]\cr
Control object.}

\item{solver}{[\code{character(1)}]\cr
Solver name, i.e., on of 2APP, 2OPT or CHRIS.}

\item{bin}{[\code{character(1)}]\cr
Full path to the binary executable.}
}
\value{
[\code{list}]
}

