% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTourLength.R
\name{computeTourLength}
\alias{computeTourLength}
\title{Compute tour length given a TSP instance and a tour/permutation.}
\usage{
computeTourLength(x, tour, close.tour = TRUE, round = FALSE)
}
\arguments{
\item{x}{[\code{Network}]\cr
TSP instance.}

\item{tour}{[\code{integer}]\cr
Permutation of nodes.}

\item{close.tour}{[\code{logical(1)}]\cr
Should the tour be closed? I.e., do we want a roundtrip tour?
Default is \code{TRUE}.}

\item{round}{[\code{logical}]\cr
Should the distances be rounded? Default is \code{FALSE}. Concorde internally
does this.}
}
\value{
[\code{numeric(1)}]
}
\description{
Compute tour length given a TSP instance and a tour/permutation.
}

