% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findShortestHamiltonianPath.R
\name{findShortestHamiltonianPath}
\alias{findShortestHamiltonianPath}
\title{Shortest Hamiltonian Path computation.}
\usage{
findShortestHamiltonianPath(x, start.id, dest.id)
}
\arguments{
\item{x}{[Network]
Network (see package netgen).}

\item{start.id}{[integer(1)]
ID of the start node.}

\item{dest.id}{[integer(1)]
ID of the end node.}
}
\value{
[integer]
  Shortest hamiltonian path.
}
\description{
Find the Shortest Hamiltonian Path between two cities by
means of a TSP formulation.
}

