% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFeatureSet.R
\name{getAvailableFeatureSets}
\alias{getAvailableFeatureSets}
\title{Available feature (sub)sets.}
\usage{
getAvailableFeatureSets()
}
\description{
May be used as black list for \code{getFeatureSet}.
}

