% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAvailableSolverNames.R
\name{getAvailableSolverNames}
\alias{getAvailableSolverNames}
\title{Returns all currently implemented solver names.}
\usage{
getAvailableSolverNames()
}
\value{
[\code{character}]
}
\description{
Returns all currently implemented solver names.
}
