% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCentroidFeatureSet.R
\name{getCentroidFeatureSet}
\alias{getCentroidFeatureSet}
\title{Feature: Centroid coordinates and statistics of distances to centroid.}
\usage{
getCentroidFeatureSet(x, include.costs = FALSE, normalize = FALSE, ...)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}

\item{normalize}{[\code{logical(1)}]\cr
Additionally calculate the normalization for the features? The default is
\code{FALSE}.}

\item{...}{[any]\cr
Not used.}
}
\value{
[\code{list}]
}
\description{
Feature: Centroid coordinates and statistics of distances to centroid.
}
