% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterFeatureSet.R
\name{getClusterFeatureSet}
\alias{getClusterFeatureSet}
\title{Feature: statistics of the minimum spanning tree edges.}
\usage{
getClusterFeatureSet(x, epsilon = 0.01, include.costs = FALSE,
  normalize = FALSE, ...)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{epsilon}{[\code{numeric(1)}]\cr
Probability for reachability computation in dbscan clustering method.
Default is 0.01.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}

\item{normalize}{[\code{logical(1)}]\cr
Should the features be normalized?
Note that normalization is not available for all features. E.g. skew, sd
and varcoeff features as well as the VRP features are not normalized.
Default is \code{FALSE}.}

\item{...}{[any]\cr
Not used.}
}
\value{
[\code{list}]
}
\description{
Feature: statistics of the minimum spanning tree edges.
}
