% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getConvexHullFeatureSet.R
\name{getConvexHullFeatureSet}
\alias{getConvexHullFeatureSet}
\title{Feature: number of points on convex hull and spanning area.}
\usage{
getConvexHullFeatureSet(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{list}]
}
\description{
Feature: number of points on convex hull and spanning area.
}

