% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureSet.R
\name{getDefaultFeatureFunArgs}
\alias{getDefaultFeatureFunArgs}
\title{Returns list of parameters defaults for feature computation.}
\usage{
getDefaultFeatureFunArgs()
}
\value{
[\code{list}]
}
\description{
Returns list of parameters defaults for feature computation.
}
