% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDistanceFeatureSet.R
\name{getDistanceFeatureSet}
\alias{getDistanceFeatureSet}
\title{Feature: Statistics of Nearest-Neighbour distances.}
\usage{
getDistanceFeatureSet(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{list}]
}
\description{
Feature: Statistics of Nearest-Neighbour distances.
}

