% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getFeatureSet.R
\name{getFeatureSet}
\alias{getFeatureSet}
\title{Feature: all features.}
\usage{
getFeatureSet(x, black.list = character(0), include.times = FALSE)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{black.list}{[\code{character}]\cr
Optional black list of feature sets which we do not want to compute. Default
is the empty character vector.}

\item{include.times}{[\code{logical(1)}]\cr
Include the times needed to compute the specific feature sets as additional
features? Default is \code{FALSE}. Time is measured via \code{proc.time}.}
}
\value{
[\code{list}]
  Named list of features.
}
\description{
Feature: all features.
}

