% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getFractionOfPointsNearBoundingBox.R
\name{getFractionOfPointsNearBoundingBox}
\alias{getFractionOfPointsNearBoundingBox}
\title{Feature: fraction of points near bounding boxplot.}
\usage{
getFractionOfPointsNearBoundingBox(x, distance_fraction = 0.1)
}
\arguments{
\item{x}{[\code{\link[netgen]{Network}}]\cr
Network.}

\item{distance_fraction}{[\code{numeric}]\cr
Fraction of distance between min and max coordinates of each dimension
which should points be allowed lie inbetween.}
}
\value{
[\code{list}]
}
\description{
Feature: fraction of points near bounding boxplot.
}

