% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHullDistanceFeatureSet.R
\name{getHullDistanceFeatureSet}
\alias{getHullDistanceFeatureSet}
\title{Feature: statistics of the distance to the hull edges}
\usage{
getHullDistanceFeatureSet(x, skewness.type = NULL, include.costs = FALSE)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{skewness.type}{[\code{integer(1)}]\cr
Definition for computing the skewness as described in \code{\link[e1071]{skewness}}.
The default is \code{type = 3}.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}
}
\value{
[\code{list}]
}
\description{
Feature: statistics of the distance to the hull edges
}
