% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getModesFeatureSet.R
\name{getModesFeatureSet}
\alias{getModesFeatureSet}
\title{Feature: modes features.}
\usage{
getModesFeatureSet(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{list}]
}
\description{
Feature: modes features.
}

