% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMonitoringFeatureSet.R
\name{getMonitoringFeatureSet}
\alias{getMonitoringFeatureSet}
\title{Feature: monitoring of algorithm runs.}
\usage{
getMonitoringFeatureSet(
  times,
  incumbents,
  lower.bound = 0,
  include.costs = FALSE,
  raw.data = FALSE,
  ...
)
}
\arguments{
\item{times}{[\code{numeric}]\cr
Times at which logging took place.}

\item{incumbents}{[\code{numeric}]\cr
Logged best value for each timing.}

\item{lower.bound}{[\code{numeric(1)}]\cr
Lower bound for the computation of the area under the curve.
Default is 0.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}

\item{raw.data}{[\code{logical(1)}]\cr
Set this to \code{TRUE} if you want the raw data computed on the
trajectory instead of the statistics.
Default is \code{FALSE}.}

\item{...}{[any]\cr
Not used.}
}
\value{
[\code{list}] Named list.
}
\description{
Feature: monitoring of algorithm runs.
}
