% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getMSTFeatureSet.R
\name{getMSTFeatureSet}
\alias{getMSTFeatureSet}
\title{Feature: statistics of the minimum spanning tree edges.}
\usage{
getMSTFeatureSet(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{list}]
}
\description{
Feature: statistics of the minimum spanning tree edges.
}

