% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getNearestNeighbourFeatureSet.R
\name{getNearestNeighbourFeatureSet}
\alias{getNearestNeighbourFeatureSet}
\title{Feature: Statistics of Nearest-Neighbour distances.}
\usage{
getNearestNeighbourFeatureSet(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{list}]
}
\description{
Feature: Statistics of Nearest-Neighbour distances.
}

