% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNNGFeatures.R
\name{getNNGFeatureSet}
\alias{getNNGFeatureSet}
\title{Feature: statistics of the nearest neighbour graph}
\usage{
getNNGFeatureSet(x, ks = NULL, include.costs = FALSE, ...)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{ks}{[\code{integer}]\cr
List of \eqn{k} values for \eqn{k}-Nearest-Neighbour-Graph computation.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}

\item{...}{[any]\cr
Not used.}
}
\value{
[\code{list}]
}
\description{
Feature: statistics of the nearest neighbour graph
}
