% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeSolver.R
\name{getSolverProperties}
\alias{getSolverProperties}
\title{Get the properties/tags of the solver.}
\usage{
getSolverProperties(solver)
}
\arguments{
\item{solver}{[\code{TSPSolver}]\cr
Solver object.}
}
\value{
[\code{character}]
}
\description{
Get the properties/tags of the solver.
}

