% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVRPFeatureSet.R
\name{getVRPFeatureSet}
\alias{getVRPFeatureSet}
\title{Feature: Statistics of distances between obligatory/optional customers and
depots.}
\usage{
getVRPFeatureSet(x, include.costs = FALSE)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}
}
\value{
[\code{list}]
}
\description{
Feature: Statistics of distances between obligatory/optional customers and
depots.
}
\note{
Only orienteering instances with two depots are supported at the moment.
}
