% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSolver.R
\name{hasProperties}
\alias{hasProperties}
\title{Check if learner has certain properties.}
\usage{
hasProperties(solver, properties)
}
\arguments{
\item{solver}{[\code{TSPSolver}]\cr
Solver object.}

\item{properties}{[\code{character}]\cr
Vector of strings.}
}
\value{
[\code{logical(1)}] \code{TRUE} if the solver has all properties, \code{FALSE} otherwise.
}

