% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSolver.R
\name{isExternalSolver}
\alias{isExternalSolver}
\title{Check if solver is called via an external binary.}
\usage{
isExternalSolver(solver)
}
\arguments{
\item{solver}{[\code{TSPSolver}]}
}
\value{
[\code{logical(1)}]
}
\description{
Check if solver is called via an external binary.
}
