% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isPermutation}
\alias{isPermutation}
\title{Check if vector contains permutation of number 1:n}
\usage{
isPermutation(x, source = seq(length(x)))
}
\arguments{
\item{x}{[\code{numeric}]\cr
Vector to check.}

\item{source}{[\code{numeric}]\cr
Vector which we want to compare with. Default is 1, ..., length(x).}
}
\value{
[\code{logical(1)}]
}
\description{
Check if vector contains permutation of number 1:n
}

