% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_feature_15_23.R
\name{ls_feature_15_23}
\alias{ls_feature_15_23}
\title{Compute features for the TSP instance using linkern algorithm. from feat: 15-23}
\arguments{
\item{solver_path}{Path to the Linkern algorithm executable.}

\item{instance_path}{Path to the TSP instance file.}
}
\value{
A vector containing computed features.
}
\description{
This function combines multiple features computed using linkern algorithm for a TSP instance.
}
\examples{
\dontrun{
  bc_features("/path/to/linkern", "/path/to/tsp_instance")
}

}
