% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSolver.R
\name{makeSolver}
\alias{makeSolver}
\title{Generator for TSP solvers.}
\usage{
makeSolver(solver)
}
\arguments{
\item{solver}{[\code{character(1)}]\cr
Solver Id.}
}
\value{
[\code{TSPSolver}]
}
\description{
Construct a TSP solver object.
}

