% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeSolver.R
\name{makeTSPSolverInternal}
\alias{makeTSPSolverInternal}
\title{Internal generator for TSP solvers.}
\usage{
makeTSPSolverInternal(cl, short.name, name, properties, par.set,
  packages = NULL, description = NULL)
}
\arguments{
\item{cl}{[\code{character(1)}]\cr
Class name for the TSP solver.}

\item{short.name}{[\code{character(1)}]\cr
Short name / identifier for the solver.}

\item{name}{[\code{character(1)}]\cr
Long name of the solver.}

\item{properties}{[\code{character}]\cr
Character vector of characterizing properties of the learner, e.g., \dQuote{external}
if the solver is called via an external script.}

\item{par.set}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
Set of solver parameters with constraints and defaults.}

\item{packages}{[\code{character}]\cr
Character vector of package names of R packages which are necessary to
run the solver.}

\item{description}{[\code{character(1)}]\cr
Optional description of the solver or solvers functioning principle.}
}
\value{
[\code{TSPSolver}]
}
\description{
Internal generator for TSP solvers.
}

