% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/measureTime.R
\name{measureTime}
\alias{measureTime}
\title{Helper for measuring time.}
\usage{
measureTime(expr, envir = parent.frame())
}
\arguments{
\item{expr}{[\code{expression}]\cr
Expression the time should be measured.}

\item{envir}{[\code{environment}]\cr
Environment in which expr should be evaluated.}
}
\value{
Return value of evaluated \code{expr} with additional attribute
  \dQuote{time_elapsed}.
}
\description{
Simple wrapper around \code{proc.time}.
}

