% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolveInstance.helpers.R
\name{mutateNormal}
\alias{mutateNormal}
\title{Local mutation operator of EA.}
\usage{
mutateNormal(coordinates, p, sigma)
}
\arguments{
\item{coordinates}{[\code{matrix}]\cr
Numeric matrix of city coordinates, rows denote cities.}

\item{p}{[\code{numeric(1)}]\cr
Mutation probability.}

\item{sigma}{[\code{numeric(1)}]\cr
Standard deviation of normal noise.}
}
\value{
[\code{matrix}]
  Numeric matrix of globally mutated city coordinates.
}
\description{
Adds some normal noise to a fraction of city coordinates.
}

