% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolveInstance.helpers.R
\name{mutateUniform}
\alias{mutateUniform}
\title{Global mutation operator for EA.}
\usage{
mutateUniform(coordinates, p)
}
\arguments{
\item{coordinates}{[\code{matrix}]\cr
Numeric matrix of city coordinates, rows denote cities.}

\item{p}{[\code{numeric(1)}]\cr
Mutation probability.}
}
\value{
[\code{matrix}]
  Numeric matrix of globally mutated city coordinates.
}
\description{
Selects a fraction of cities at reandom and replaces these by new cities
generated uniform at random in [0,1] x [0,1].
}
