% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{netgenToTSPmeta}
\alias{netgenToTSPmeta}
\title{Convert netgen object to tspmeta TSP instance.}
\usage{
netgenToTSPmeta(x)
}
\arguments{
\item{x}{[Network]
Network.}
}
\value{
[tsp_instance]
}
\description{
Convert netgen object to tspmeta TSP instance.
}

