% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeSolver.R
\name{prepareInstance}
\alias{prepareInstance}
\title{Preprocess instance before running the solver on it.}
\usage{
prepareInstance(solver, instance)
}
\arguments{
\item{solver}{[\code{TSPSolver}]\cr
Solver object.}

\item{instance}{[\code{Network}]\cr
Instance to solver.}
}
\value{
[any] Specific to the solver.
}
\description{
Preprocess instance before running the solver on it.
}

