% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTSPlibTOURFile.R
\name{readTSPlibTOURFile}
\alias{readTSPlibTOURFile}
\title{Extract tour from TSPlib tour file.}
\usage{
readTSPlibTOURFile(file.path)
}
\arguments{
\item{file.path}{[\code{character(1)}]\cr
Path to TSPlib tour file.}
}
\value{
[\code{list}] List with components \code{tour} and \code{tour.length}.
}
\description{
Extract tour from TSPlib tour file.
}
