% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{removeFileExtension}
\alias{removeFileExtension}
\title{Removes the file extension of a given filename.}
\usage{
removeFileExtension(filename)
}
\arguments{
\item{filename}{[character(1)]
Filename.}
}
\value{
Truncated filename, i.e., filename without last file extension.
}
\description{
Removes the file extension of a given filename.
}
