% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolveInstance.helpers.R
\name{rescaleCoordinates}
\alias{rescaleCoordinates}
\title{Rescale city coordinates.}
\usage{
rescaleCoordinates(coordinates)
}
\arguments{
\item{coordinates}{[\code{matrix}]\cr
Numeric matrix of city coordinates, rows denote cities.}
}
\description{
Rescale coordinates to [0,1] x [0,1].
}
