% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolveInstance.helpers.R
\name{roundToGrid}
\alias{roundToGrid}
\title{Round instance (points are placed in the center of the grids).}
\usage{
roundToGrid(coordinates, n = 100L)
}
\arguments{
\item{coordinates}{[\code{matrix}]\cr
Numeric matrix of city coordinates, rows denote cities.}

\item{n}{[\code{numeric(1)}]\cr
Number of cells desired, i.e., grid resolution.}
}
\value{
[\code{matrix}]
  Numeric matrix of rounded city coordinates.
}
\description{
Round instance (points are placed in the center of the grids).
}
