% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.concorde_linkern.R
\name{run.concorde_linkern}
\alias{run.concorde_linkern}
\title{Runs the Concorde or Linkern algorithm and returns the output.}
\usage{
\method{run}{concorde_linkern}(solver_path, method, instance_path, args = "", concorde_runtime = 0)
}
\arguments{
\item{solver_path}{[\code{character(1)}]\cr
Path to the Concorde or Linkern solver executable.}

\item{method}{[\code{character(1)}]\cr
Method to use (e.g., "concorde" or "linkern").}

\item{instance_path}{[\code{character(1)}]\cr
Path to the instance file to be solved.}

\item{args}{[\code{character(1)}]\cr
Additional command-line arguments for the solver. Default is an empty string.}

\item{concorde_runtime}{[\code{numeric(1)}]\cr
Maximum runtime limit for Concorde algorithm (in seconds). Default is 0 (no limit).}
}
\value{
result [\code{character(n)}]\cr
  Output of the Concorde or Linkern algorithm.
}
\description{
Runs the Concorde or Linkern algorithm and returns the output.
}
