% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.lkh.R
\name{run.lkh}
\alias{run.lkh}
\title{Solver: LKH}
\usage{
\method{run}{lkh}(solver, instance, runs = 1L, seed = as.integer(runif(1L) *
  2^15), cutoff.time = 0, opt.tour.length = NULL, max.trials = 1000L,
  full.matrix = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{solver}{[\code{TSPSolver}]\cr
TSPSolver object.}

\item{instance}{[\code{character(1)} | \code{Network}]\cr
Path to TSP instance in \href{https://www.iwr.uni-heidelberg.de/groups/comopt/software/TSPLIB95/}{TSPlib}
format or \code{Network} object.}

\item{runs}{[\code{integer(1)}]\cr
Number of independent runs. This value is fixed to 1.}

\item{seed}{[\code{integer(1)}]\cr
Seed for the random numbers generator.
Defaults to a random integer between 1 and \eqn{2^15}.}

\item{cutoff.time}{[\code{integer(1)}]\cr
Maximal running time in seconds.
Default is 0, i.e., no cutoff time.}

\item{opt.tour.length}{[\code{integer(1) | NULL}]\cr
Length of an optimal TSP tour. This is only relevant for benchmarking.
Keep in mind, that in case of a given optimal tour length most internal
stopping criteria of EAX are deacitvated. The algorithm stops if an optimal
tour is found or the cutoff time is reached.}

\item{max.trials}{[\code{integer(1)}]\cr
Maximal number of iterations.
Default is 1000.}

\item{full.matrix}{[\code{logical(1)}]\cr
Only relevant if \code{instance} is a \code{Network}.
If \code{TRUE}, the network is exported to TSPlib format
with explicit edge weight definition.
Default is \code{FALSE}.}

\item{verbose}{[\code{logical(1)}]\cr
Should the output of the solver be printed?
Default is \code{FALSE}.}

\item{...}{[any]\cr
Not used at the moment.}
}
\value{
[\code{\link{TSPSolverResult}}]
}
\description{
Inexact TSP solvers based the Lin-Kernigham heuristic.
}
\note{
This solver requires integer inter-city distances.
}
\references{
Helsgaun, K. (2000). An effective implementation of the lin-kernighan traveling
salesman heuristic. European Journal of Operational Research, 126:106-130.

Helsgaun, K. (2009). General k-opt submoves for the Lin-Kernighan TSP heuristic.
Mathematical Programming Computation, 1(2-3):119-163.
}
