% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runSolver.R
\name{runSolver}
\alias{runSolver}
\title{Apply solver to instance.}
\usage{
runSolver(solver, instance, solver.pars, ...)
}
\arguments{
\item{solver}{[\code{character(1)} || \code{TSPSolver}]\cr
Solver.}

\item{instance}{[\code{Network}]\cr
Instance to solve.}

\item{solver.pars}{[\code{list}]\cr
Named list of parameters for the solver.}

\item{...}{[any]
Parameters for the solver.}
}
\value{
[\code{TSPSolverResult}]
}
\description{
This is a central function. It expects a solver (either a characterizing string
  or a TSPSolver object), a Network to operate on and optional parameter for the
  algorithm/solver. It then runs the solver on the instance and returns a solver
  result object.
}

