% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adapters.R
\name{runSolverFromTSPPackage}
\alias{runSolverFromTSPPackage}
\title{Adapter/interface to the \code{\link[TSP]{solve_TSP}} method.}
\usage{
runSolverFromTSPPackage(solver, instance, solver.pars = NULL)
}
\arguments{
\item{solver}{[\code{character(1)}]\cr
Name of the solver.}

\item{instance}{[\code{Network}]\cr
Netgen network.}

\item{solver.pars}{[\code{list}]\cr
Named list of solver parameters.}
}
\value{
[\code{list}]
}
\description{
Adapter/interface to the \code{\link[TSP]{solve_TSP}} method.
}

