% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStatisticsOnNumericVector.R
\name{skew}
\alias{skew}
\title{Computes skewness of a numeric vector (analogous to skewness type 3 of e1071)}
\usage{
skew(x)
}
\arguments{
\item{x}{[\code{numeric}]\cr
Vector used for computing the skewness.}
}
\value{
[\code{numeric(1)}]
}
\description{
Computes skewness of a numeric vector (analogous to skewness type 3 of e1071)
}
