% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAngleFeatureSet.R
\name{getAngleFeatureSet}
\alias{getAngleFeatureSet}
\title{Feature: statistics of angles between nodes and their two nearest neighbors.}
\usage{
getAngleFeatureSet(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{list}]
}
\description{
Feature: statistics of angles between nodes and their two nearest neighbors.
}

