% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAngleFeatureSet.R
\name{getAngleFeatureSet}
\alias{getAngleFeatureSet}
\title{Feature: statistics of angles between nodes and their two nearest neighbors.}
\usage{
getAngleFeatureSet(x, drop.duplicates = FALSE, include.costs = FALSE)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{drop.duplicates}{[\code{logical(1)}]\cr
Should duplicated node coordinates be dropped?
Default is \code{FALSE}.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}
}
\value{
[\code{list}]
}
\description{
Feature: statistics of angles between nodes and their two nearest neighbors.
}
