% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundingBoxFeatureSet.R
\name{getBoundingBoxFeatureSet}
\alias{getBoundingBoxFeatureSet}
\title{Feature: fraction of points near bounding boxplot.}
\usage{
getBoundingBoxFeatureSet(
  x,
  distance_fraction = 0.1,
  include.costs = FALSE,
  normalize = FALSE,
  ...
)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{distance_fraction}{[\code{numeric}]\cr
Fraction of distance between min and max coordinates of each dimension
which should points be allowed lie inbetween.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}

\item{normalize}{[\code{logical(1)}]\cr
Additionally calculate the normalization for the features? The default is
\code{FALSE}.}

\item{...}{[any]\cr
Not used.}
}
\value{
[\code{list}]
}
\description{
Feature: fraction of points near bounding boxplot.
}
