% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getCentroidFeatureSet.R
\name{getCentroidFeatureSet}
\alias{getCentroidFeatureSet}
\title{Feature: Centroid coordinates and statistics of distances to centroid.}
\usage{
getCentroidFeatureSet(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{list}]
}
\description{
Feature: Centroid coordinates and statistics of distances to centroid.
}

