% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConvexHullFeatureSet.R
\name{getConvexHullFeatureSet}
\alias{getConvexHullFeatureSet}
\title{Feature: number of points on convex hull and spanning area.}
\usage{
getConvexHullFeatureSet(x, feature.set = NULL, include.costs = FALSE,
  normalize = FALSE, ...)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{feature.set}{[\code{character}]\cr
Subset of hull feature sets that should be computed. Possible choices are
\code{"points"} (ratio of points defining the convex hull), \code{"area"}
(area of the convex hull), \code{"edges"} (summary statistics of the
lengths of the hull edges) and \code{"dists"} (summary statistics of the
distances of each city to its closest edge of the hull).
Per default (\code{NULL}), all of the four previously listed feature sets will
be computed.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}

\item{normalize}{[\code{logical(1)}]\cr
Should the features be normalized?
Note that normalization is not available for all features. E.g. skew, sd
and varcoeff features as well as the VRP features are not normalized.
Default is \code{FALSE}.}

\item{...}{[any]\cr
Not used.}
}
\value{
[\code{list}]
}
\description{
Feature: number of points on convex hull and spanning area.
}
