% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConvexHullFeatureSet.R
\name{getConvexHullFeatureSet}
\alias{getConvexHullFeatureSet}
\title{Feature: number of points on convex hull and spanning area.}
\usage{
getConvexHullFeatureSet(x, include.costs = FALSE)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}
}
\value{
[\code{list}]
}
\description{
Feature: number of points on convex hull and spanning area.
}
