% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDistanceFeatureSet.R
\name{getDistanceFeatureSet}
\alias{getDistanceFeatureSet}
\title{Feature: Statistics of Nearest-Neighbour distances.}
\usage{
getDistanceFeatureSet(x, include.costs = FALSE)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}
}
\value{
[\code{list}]
}
\description{
Feature: Statistics of Nearest-Neighbour distances.
}

