% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureSet.R
\name{getFeatureSet}
\alias{getFeatureSet}
\title{Feature: all features.}
\usage{
getFeatureSet(x, black.list = character(0), include.costs = FALSE,
  feature.fun.args = getDefaultFeatureFunArgs())
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{black.list}{[\code{character}]\cr
Optional black list of feature sets which we do not want to compute. Default
is the empty character vector.}

\item{include.costs}{[\code{logical(1)}]\cr
Include the times needed to compute the specific feature sets as additional
features? Default is \code{FALSE}. Time is measured via \code{proc.time}.}

\item{feature.fun.args}{[\code{list}]\cr
List of lists. Each component of the list corresponds to one of the feature sets
and the subordered list contains vectors of values for each parameter. Name
the sublists according to \link{getAvailableFeatureSets}.}
}
\value{
[\code{list}]
  Named list of features.
}
\description{
Feature: all features.
}
\examples{
  library(netgen)
  x = generateRandomNetwork(n.points = 50L)
  # get all features available in salesperson with default parameters
  # Note that we always exclude the VRP problem specific VRP features, since
  # we are operating on a simple euclidean TSP instance without depots and
  # arrival times.
  fs = getFeatureSet(x, black.list = "VRP")

  # now include the costs of computing (in seconds) for each set of parameters
  # as additional features
  fs = getFeatureSet(x, black.list = "VRP", include.costs = TRUE)

  # compute all but MST and distance features
  fs = getFeatureSet(x, black.list = c("VRP", "MST", "Distance"))

  # now set user-defined parameter values for cluster
  args = list("Cluster" = list("epsilon" = c(0.01, 0.05, 0.1, 0.2, 0.3)))
  fs = getFeatureSet(x, black.list = "VRP", feature.fun.args = args)
}

