% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getModesFeatureSet.R
\name{getModesFeatureSet}
\alias{getModesFeatureSet}
\title{Feature: modes features.}
\usage{
getModesFeatureSet(x, include.costs = FALSE)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}
}
\value{
[\code{list}]
}
\description{
Feature: modes features.
}

