% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMSTFeatureSet.R
\name{getMSTFeatureSet}
\alias{getMSTFeatureSet}
\title{Feature: statistics of the minimum spanning tree edges.}
\usage{
getMSTFeatureSet(x, include.costs = FALSE)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{include.costs}{[\code{logical(1)}]\cr
Should costs be included as an additional feature? Default is \code{FALSE}.}
}
\value{
[\code{list}]
}
\description{
Feature: statistics of the minimum spanning tree edges.
}
