% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{readTSPlibTOURFile}
\alias{readTSPlibTOURFile}
\title{Extracts tour from TSPlib file.}
\usage{
readTSPlibTOURFile(file.path)
}
\arguments{
\item{file.path}{[\code{character(1)}]\cr
Path to TSPlib tour file.}
}
\value{
[\code{integer}]
  Permutation of the nodes.
}
\description{
Since we need only the tour and the tour length, this functions skips the
specification part.
}

