% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.lkh.R
\name{run.lkh}
\alias{run.lkh}
\title{Solver: LKH}
\usage{
\method{run}{lkh}(
  solver,
  instance,
  seed = as.integer(ceiling(runif(1L) * 2^15)),
  cutoff.time = NULL,
  opt.tour.length = NULL,
  max.trials = NULL,
  with.restarts = FALSE,
  full.matrix = FALSE,
  verbose = FALSE,
  log.trajectory = TRUE,
  work.dir = NULL,
  output.files.prefix = NULL,
  keep.output.files = FALSE,
  more.args = list(),
  ...
)
}
\arguments{
\item{solver}{[\code{TSPSolver}]\cr
TSPSolver object.}

\item{instance}{[\code{character(1)} | \code{Network}]\cr
Path to TSP instance in \href{https://www.iwr.uni-heidelberg.de/groups/comopt/software/TSPLIB95/}{TSPlib}
format or \code{Network} object.}

\item{seed}{[\code{integer(1)}]\cr
Seed for the random numbers generator.
Defaults to a random integer between 1 and \eqn{2^15}.}

\item{cutoff.time}{[\code{integer(1)}]\cr
Maximal running time in seconds.
Default is NULL, i.e., no cutoff time.}

\item{opt.tour.length}{[\code{integer(1) | NULL}]\cr
Length of an optimal TSP tour. This is only relevant for benchmarking.
Keep in mind, that in case of a given optimal tour length most internal
stopping criteria of EAX are deacitvated. The algorithm stops if an optimal
tour is found or the cutoff time is reached.}

\item{max.trials}{[\code{integer(1)}]\cr
Maximal number of iterations.
Default is the number of nodes of the instance.}

\item{with.restarts}{[\code{logical(1)}]\cr
Should LKH restart if a plateau is reached?
Default is \code{FALSE}.}

\item{full.matrix}{[\code{logical(1)}]\cr
Only relevant if \code{instance} is a \code{Network}.
If \code{TRUE}, the network is exported to TSPlib format
with explicit edge weight definition.
Default is \code{FALSE}.}

\item{verbose}{[\code{logical(1)}]\cr
Should the output of the solver be printed?
Default is \code{FALSE}.}

\item{log.trajectory}{[\code{logical(1)}]\cr
Should the optimization progress be logged?
Default is \code{FALSE}.}

\item{work.dir}{[\code{character(1)}]\cr
Path to working directory of the solver.
Defaults to temporary directory via \code{tempfile()}.}

\item{output.files.prefix}{[\code{character(1)}]\cr
Prefix for file names used by solver for storing output.
Default is a randomly generated temporary prefix.}

\item{keep.output.files}{[\code{logical(1)}]\cr
Flag indicating whether output files shall be kept or deleted.
Default is \code{FALSE}, i.e., output files are processed and
output is loaded into R object. Afterwards the output files are
deleted. This flag is particularly useful in large benchmark studies if
\code{work.dir} and \code{output.files.prefix} is set.}

\item{more.args}{[\code{list}]\cr
Named list of parameter which shall be written to the LKH parameter file.
Note that 1) the names should be all uppercase and 2) there is no argument
check. Default ist the empty list.}

\item{...}{[any]\cr
Not used at the moment.}
}
\value{
[\code{\link{TSPSolverResult}}]
}
\description{
Inexact TSP solvers based the Lin-Kernigham heuristic.
}
\note{
This solver requires integer inter-city distances.
}
\references{
Helsgaun, K. (2000). An effective implementation of the lin-kernighan traveling
salesman heuristic. European Journal of Operational Research, 126:106-130.

Helsgaun, K. (2009). General k-opt submoves for the Lin-Kernighan TSP heuristic.
Mathematical Programming Computation, 1(2-3):119-163.
}
