% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSolver.R
\name{run}
\alias{run}
\title{Run a TSP solver to solve a problem instance.}
\usage{
run(solver, instance, solver.pars, ...)
}
\arguments{
\item{solver}{[\code{TSPSolver}]\cr
Solver object.}

\item{instance}{[\code{Network}]\cr
TSP instance to solve.}

\item{solver.pars}{[\code{list}]\cr
Potentially empty named list of solver parameters passed to the solver.}

\item{...}{[any]\cr
Not used.}
}
\value{
[\code{TSPSolverResult}]
}
\description{
Run a TSP solver to solve a problem instance.
}
